condexpr pointer casting, null pointer casting, some testing
When each branch of a condexpr is a pointer, resolve the result to the
appropriate type.

Add an implicit cast from literal 0 to null pointer. Applies to
assignments / conditional operator; still needs to be done for initializers

Adds some for type resolution; I tried to structure in a way that's
not too dependent on the final AST - it compares the dumped type
of a statement to an expected string.