Source: require source files to be utf-8 encoded
Currently this just scans the whole file up-front to ensure that it is
utf-8 encoded. The scan is performed in preprocessor even though addSource
seems like a more natural fit because returning an error from that function
invalidates the source and path buffers which are needed for error messages.

Subsequent tokenizer enhancements will be needed for consuming unicode
codepoints instead of bytes (e.g. for identifiers); at that time the extra
scan can be removed and tokenizer.next() can either return an error or a special
token to indicate invalid utf8