Parser: correctly parse c2x attributes after declarator
This fixes the immediate issue in #219, but since `directDeclarator` can now
produce attributes, each place that calls it should be audited to make sure
that it properly adds the attributes if necessary. However I'm not sure what
the relevant C code would look like to exercise it in a test.

Fixes #219