Source: move invalid-utf8 location storage into Compilation
Store invalid UTF-8 locations in a hash map in Compilation, instead of directly
in Source. This incurs a fixed cost of 16 bytes per Compilation to store the
hash map, but saves 12 bytes per Source which is valid UTF-8. Since there will
generally be many Sources, and invalid UTF-8 is an uncommon and fatal error,
this should usually result in less memory usage.