Preprocessor: handle macros which expand to `defined`
Since `defined` needs to look at its argument before expansion
happens, it needs to be special cased in expandMacroExhaustive.

It can't be a special function-like macro because that will cause
incorrect expansion outside of preprocessor expression contexts.
It can't be handled solely in `Preprocessor.expr` because the
argument will have been expanded by that point.

Closes #337