Preprocessor: track nonreplaced function-like macros
The text of 6.10.3.4.2 is somewhat ambiguous:

    These nonreplaced macro name preprocessing tokens are no longer available
    for further replacement even if they are later (re)examined in contexts in
    which that macro name preprocessing token would otherwise have been
    replaced.

The immediately preceding text is describing the rule for not recursively
replacing a macro with itself if it appears in its own replacement list,
but a token is also "nonreplaced" if it is a function-like macro that is
not followed by a left parenthesis. GCC and clang both do not expand
a macro in this case during a rescan.

This rule only applies if a macro is nonexpanded in a replacement list;
it does not apply to argument pre-expansion.

Closes #339