all: fix unnecessary uses of 'var'
These were noticed downstream, since I am in the process of implementing
an error for uses of 'var' when one provably could have used 'const'.

Note that the changes in 'deps' are not currently in the Zig codebase,
but will be a part of an imminent PR there. I thought it made sense to
make them now to avoid having to sync this code once the new error is
introduced.