Parser: Don't create implicit functions during record parsing
If an implicit function def was created during record parsing, it was
possible to have the parser decls and record fields get out of sync, so that
there were more decls than fields (because the implicit function def would
be treated as a record field declarator). This could cause a crash if
subsequent fields contained attributes, because we store attributes based
on the number of fields and not the number of decls. Thus, we would store
e.g. only 2 sets of field attributes even though there were 3 decls.

Fixes #639