Tree: represent nodes as tagged unions with a packed representation
Working with the current AST is annoying and difficult
with similar constructs having multiple representations
and dealing with multiarray list access and `@intFromEnum`
and different data being stored in different data structures.

The goal of this branch is to simplify the usage by having
a tagged union containing all the information of that node.