add `-resource-dir <dir>` CLI option
This allows overriding the path to the "resource directory", which is
the *parent* of Aro's builtin 'include' directory. This is useful when
invoking Aro programmatically from the Zig build system since it avoids
needing to copy everything into an installation prefix.

The option mimics Clang's `-resource-dir` option in form and function.