Preprocessor: Add implicit includes
Implicit includes are included at the beginning of the primary source file
(using the -include flag in the driver)

The first directory searched for <file> is the Driver's working directory
instead of the directory containing the main source file. If not found there,
it is searched for in the remainder of the #include "..." search chain as
normal.